+++ src/main/java/kitchenpos/order/application/TableService.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.application;
 
-import kitchenpos.order.domain.OrderDao;
-import kitchenpos.order.domain.OrderTableDao;
+import kitchenpos.order.dao.OrderDao;
+import kitchenpos.order.dao.OrderTableDao;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderTable;
 import org.springframework.stereotype.Service;
