+++ src/main/java/kitchenpos/order/dao/JdbcTemplateOrderTableDao.java
@@ -1,5 +1,6 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.dao;
 
+import kitchenpos.order.domain.OrderTable;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
