+++ src/main/java/kitchenpos/order/dao/OrderLineItemDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.dao;
 
 import kitchenpos.order.domain.OrderLineItem;
